+++ packages/edit-navigation/src/components/block-placeholder/menu-items-to-blocks.js
@@ -11,7 +11,7 @@ import { createBlock } from '@wordpress/blocks';
 /**
  * Internal dependencies
  */
-import { menuItemToBlockAttributes } from '../../store/utils';
+import { menuItemToBlockAttributes } from '../../store/transform';
 
 /**
  * Convert a flat menu item structure to a nested blocks structure.