+++ packages/react-native-editor/CHANGELOG.md
@@ -10,12 +10,19 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+
+## 1.62.0
 -   [**] [Embed block] Implement WP embed preview component [#34004]
 -   [*] [Embed block] Fix content disappearing on Android when switching light/dark mode [#34207]
 -   [*] Embed block: Add device's locale to preview content [#33858]
 -   [**] Fix Android-only issue of main toolbar initial position being wrong when RTL [#34617]
 -   [**] Embed block: Implemented the No Preview UI when an embed is successful, but we're unable to show an inline preview [#34626]
 -   [*] Column block: Translate column width's control labels [#34777]
+-   [**] Enable embed preview for Instagram and Vimeo providers. [#34563]
+-   [**] Embed block: Add error bottom sheet with retry and convert to link actions. [#34604]
+
+## 1.61.1
+-   [*] Fix crash related to reusable blocks in the block picker. [#34873]
 
 ## 1.61.0
 -   [**] Enable embed preview for a list of providers (for now only YouTube and Twitter) [#34446]